refactor(example): rename turnkey react example
