+++ examples/react-turnkey-wallet/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "turnkey-community-example",
+  "name": "react-turnkey-wallet-example",
   "version": "0.1.0",
   "private": true,
   "scripts": {
