+++ examples/react-turnkey-wallet/src/components/react-turnkey-wallet/transfer-card.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { useConfidentialTransfer } from "@zama-fhe/react-sdk";
 import { isAddress, type Address } from "viem";
-import { parseAmountSafe } from "@/lib/turnkey-community-example/utils";
+import { parseAmountSafe } from "@/lib/react-turnkey-wallet/utils";
 import { MutationStatus } from "./mutation-status";
 
 export function TransferCard({
