+++ examples/react-turnkey-wallet/src/components/react-turnkey-wallet/unshield-card.tsx
@@ -7,7 +7,7 @@ import {
   indexedDBStorage,
 } from "@zama-fhe/react-sdk";
 import type { Address, Hex } from "viem";
-import { parseAmountSafe, shortAddr, txLink } from "@/lib/turnkey-community-example/utils";
+import { parseAmountSafe, shortAddr, txLink } from "@/lib/react-turnkey-wallet/utils";
 import { MutationStatus } from "./mutation-status";
 
 export function UnshieldCard({
