+++ examples/react-turnkey-wallet/src/hooks/react-turnkey-wallet/use-eth-balance.ts
